requirements.txt
aiogram==2.25.1
peewee==3.16.2
python-dotenv==1.0.0
asyncio==3.4.3
flask==2.3.2